Optimized the code
Made the code more clean and understandable. Also made an illuminatable
material, which allo ws for a full day and night cycle (once realistic
colors are chosen). The only thing left to do is to change the
background-meshes to actual objects so they can be assigned with the
illuminable material.